+++ lib/fabric/src/main/java/dev/matthiesen/fabric/matthiesen_lib/platform/MatthiesenLibFabricEmbersMessagingPlatformService.java
@@ -19,14 +19,14 @@ public class MatthiesenLibFabricEmbersMessagingPlatformService implements Matthi
     public MatthiesenLibFabricEmbersMessagingPlatformService() {}
 
     @Override
-    public void sendMessage(ServerPlayer player, Component message, float duration) {
+    public void sendMessage(ServerPlayer player, Component message, int duration) {
         if (MatthiesenLib.isModLoaded(MatthiesenLibBuiltInTextParsers.EMBER.getName())) {
             NetworkHelper.getInstance().sendMessage(player, new ImmersiveMessage(message, duration));
         }
     }
 
     @Override
-    public void sendMessage(ServerPlayer player, Component message, float duration, MatthiesenLibImmersiveMessageBuilder builder) {
+    public void sendMessage(ServerPlayer player, Component message, int duration, MatthiesenLibImmersiveMessageBuilder builder) {
         if (MatthiesenLib.isModLoaded(MatthiesenLibBuiltInTextParsers.EMBER.getName())) {
             ImmersiveMessage msg = new ImmersiveMessage(message, duration);
             if (builder != null) {
@@ -37,14 +37,14 @@ public class MatthiesenLibFabricEmbersMessagingPlatformService implements Matthi
     }
 
     @Override
-    public void sendMessage(ServerPlayer player, String message, float duration) {
+    public void sendMessage(ServerPlayer player, String message, int duration) {
         if (MatthiesenLib.isModLoaded(MatthiesenLibBuiltInTextParsers.EMBER.getName())) {
             NetworkHelper.getInstance().sendMessage(player, ImmersiveMessage.fromMarkup(duration, message));
         }
     }
 
     @Override
-    public void sendMessage(ServerPlayer player, String message, float duration, MatthiesenLibImmersiveMessageBuilder builder) {
+    public void sendMessage(ServerPlayer player, String message, int duration, MatthiesenLibImmersiveMessageBuilder builder) {
         if (MatthiesenLib.isModLoaded(MatthiesenLibBuiltInTextParsers.EMBER.getName())) {
             ImmersiveMessage msg = ImmersiveMessage.builder(duration, message);
             if (builder != null) {
@@ -55,14 +55,14 @@ public class MatthiesenLibFabricEmbersMessagingPlatformService implements Matthi
     }
 
     @Override
-    public void sendUpdateMessage(ServerPlayer player, String messageId, Component message, float duration) {
+    public void sendUpdateMessage(ServerPlayer player, String messageId, Component message, int duration) {
         if (MatthiesenLib.isModLoaded(MatthiesenLibBuiltInTextParsers.EMBER.getName())) {
             NetworkHelper.getInstance().sendUpdateMessage(player, messageId, new ImmersiveMessage(message, duration));
         }
     }
 
     @Override
-    public void sendUpdateMessage(ServerPlayer player, String messageId, Component message, float duration, MatthiesenLibImmersiveMessageBuilder builder) {
+    public void sendUpdateMessage(ServerPlayer player, String messageId, Component message, int duration, MatthiesenLibImmersiveMessageBuilder builder) {
         if (MatthiesenLib.isModLoaded(MatthiesenLibBuiltInTextParsers.EMBER.getName())) {
             ImmersiveMessage msg = new ImmersiveMessage(message, duration);
             if (builder != null) {
@@ -73,14 +73,14 @@ public class MatthiesenLibFabricEmbersMessagingPlatformService implements Matthi
     }
 
     @Override
-    public void sendUpdateMessage(ServerPlayer player, String messageId, String message, float duration) {
+    public void sendUpdateMessage(ServerPlayer player, String messageId, String message, int duration) {
         if (MatthiesenLib.isModLoaded(MatthiesenLibBuiltInTextParsers.EMBER.getName())) {
             NetworkHelper.getInstance().sendUpdateMessage(player, messageId, ImmersiveMessage.fromMarkup(duration, message));
         }
     }
 
     @Override
-    public void sendUpdateMessage(ServerPlayer player, String messageId, String message, float duration, MatthiesenLibImmersiveMessageBuilder builder) {
+    public void sendUpdateMessage(ServerPlayer player, String messageId, String message, int duration, MatthiesenLibImmersiveMessageBuilder builder) {
         if (MatthiesenLib.isModLoaded(MatthiesenLibBuiltInTextParsers.EMBER.getName())) {
             ImmersiveMessage msg = ImmersiveMessage.builder(duration, message);
             if (builder != null) {
